修正 1/20 model錯誤:Column 'ChatroomID' cannot be null

+++ backend/chat/src/main/java/com/scott/chat/controller/MemberController.java
@@ -21,7 +21,7 @@ public class MemberController {
     private MemberRepository memberRepository;
 
     @GetMapping("/{id}")
-    public ResponseEntity<Map<String, Object>> getMember(@PathVariable Long id) {
+    public ResponseEntity<Map<String, Object>> getMember(@PathVariable Integer id) {
         return memberRepository.findById(id)
             .map(member -> {
                 Map<String, Object> response = new HashMap<>();
